+++ backend/src/test/java/botobo/core/infrastructure/FileNameGeneratorTest.java
@@ -1,6 +1,7 @@
 package botobo.core.infrastructure;
 
 import botobo.core.exception.user.s3.ImageExtensionNotAllowedException;
+import botobo.core.infrastructure.s3.FileNameGenerator;
 import botobo.core.utils.FileFactory;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
